requirements.txt
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.42.0
pytest
importlib
